redux.classrooms now redux.teacher.classrooms... but we have bugs with students.
