+++ src/presentational/ClassroomsOverview.jsx
@@ -18,6 +18,7 @@ const ClassroomsOverview = (props, context) => (
                 <i className="fa fa-institution fa-4x"></i>
               </div>
               <div className="col-xs-9 text-right">
+                {console.log('!'.repeat(80), '\n', context)}
                 {(context.classrooms.loading === false) ?
                 <h1>{context.classrooms.data.length}</h1>
                 : <Spinner/>}
