+++ src/reducers/index.js
@@ -1,9 +1,9 @@
 import { combineReducers } from 'redux';
 
-import * as classrooms from './classrooms';
+import * as teacher from './teacher';
 import * as student from './student';
 import * as login from './login';
 import * as users from './users';
 
-const reducers = Object.assign({}, classrooms, student, login, users);
+const reducers = Object.assign({}, teacher, student, login, users);
 export default combineReducers(reducers);
